test: add test coverage to fs.promises.chown and filehandle.chown
